[NEW] Restored complete homepage & /blog/
Notes
=====

The issues with root-relative URLs are solved now that the website is published at `http://www.ditecco.me` instead of `https://aditecco.gitlab.io/ditecco.me/` -- I set a DNS `A` record on the domain, pointing to GL pages' IP address.
So I'm restoring the homepage and the /blog/ section, which I completely removed in a prior commit.