[NEW] Created 1st draft of 'directory index' page --notes
Notes
=====

The page is intended as a universal layout for sections of the site that need an index, such as:
- /archive/
- /blog/
- /projects/

In the case of /blog/, we can dynamically populate the item list via Jekyll's for-in loop; I'll have to think about how to implement this for the other sections.

This layout replaces the prev `dtc-blog-main.scss`.