[NEW] Rebuilt 'archived notice' as more universal 'common notice' include, w/ new stylesheet --notes
Notes
=====

'archived notice' was a brief notice displayed on websites in the /archive/ section; it was implemented as a portion of HTML directly added to the website's layout, with its own stylesheet.

This new version, implemented as a Jekyll include, is decoupled from the message or purpose of the notice; it's a universal notice component, and can be populated via front-matter.